+++ packages/react-native-sdk/src/components/Call/CallLayout/CallParticipantsGrid.tsx
@@ -52,11 +52,16 @@ export const CallParticipantsGrid = ({
   const {
     theme: { colors, callParticipantsGrid },
   } = useTheme();
-  const { useRemoteParticipants, useParticipants, useLocalParticipant } =
-    useCallStateHooks();
+  const {
+    useRemoteParticipants,
+    useParticipants,
+    useLocalParticipant,
+    useDominantSpeaker,
+  } = useCallStateHooks();
   const _remoteParticipants = useRemoteParticipants();
   const localParticipant = useLocalParticipant();
   const _allParticipants = useParticipants();
+  const dominantSpeaker = useDominantSpeaker();
   // we debounce the participants arrays to avoid unnecessary rerenders that happen when participant tracks are all subscribed simultaneously
   const remoteParticipants = useDebouncedValue(_remoteParticipants, 300);
   const allParticipants = useDebouncedValue(_allParticipants, 300);
@@ -80,7 +85,7 @@ export const CallParticipantsGrid = ({
   if (isInPiPMode) {
     participants =
       remoteParticipants.length > 0
-        ? [remoteParticipants[0] as StreamVideoParticipant]
+        ? [dominantSpeaker || (remoteParticipants[0] as StreamVideoParticipant)]
         : localParticipant
           ? [localParticipant]
           : [];